requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.9
dj-database-url==0.5.0
Django==3.2.9
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.3
imageio==2.11.1
MyQR==2.3.1
numpy==1.21.4
Pillow==8.4.0
psycopg2==2.9.2
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
typing_extensions==4.0.0
urllib3==1.26.7
whitenoise==5.3.0
